chore: Moved jwt file to openedx.core.lib
