+++ requirements/edx/doc.txt
@@ -1527,7 +1527,7 @@ uritemplate==4.1.1
     #   drf-spectacular
     #   drf-yasg
     #   google-api-python-client
-urllib3==1.26.20
+urllib3==2.2.3
     # via
     #   -c requirements/edx/../common_constraints.txt
     #   -r requirements/edx/base.txt
